+++ glasgow2024/services.py
@@ -1,5 +1,6 @@
 import svcs
-from nomnom.convention import ConventionConfiguration, ConventionTheme
+from nomnom.convention import ConventionConfiguration, ConventionTheme, HugoAwards
+from wsfs.rules import constitution_2023
 
 from . import convention
 
@@ -7,3 +8,4 @@ from . import convention
 def svcs_init(registry: svcs.Registry) -> None:
     registry.register_value(ConventionConfiguration, convention.convention)
     registry.register_value(ConventionTheme, convention.theme)
+    registry.register_value(HugoAwards, constitution_2023.hugo_awards)